Day4/pyproject.toml
[project]
name = "llm-day4"
version = "1.0.0"
description = "SK-RAG-AI Agent 교육 - Day 4"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.5.0",
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph-cli[inmem]>=0.4.0",
    "tavily-python>=0.3.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "jupyter>=1.1.0",
    "ipykernel>=6.29.0",
    "typing-extensions>=4.5.0",
]
